Avoid submitting commands when device is in VGA mode.
When someone runs "init 3", the device will be switched into VGA mode
with no notification to the driver. However, if commands are submitted
when the device is in VGA mode, they will be queued up, and then, the
next time a mode set set, an assertion in the device will take down
the entire virtual machine.

So this patch checks whether the device is in VGA mode before
submitting each command. This is racy because the switch to VGA could
happen just after we check, but there doesn't seem to be a better way.

(Apart from removing the assert() in the QXL device itself thereby
making it robust against command submission in VGA mode. But the
author of the device is opposed to that).