Keep track of bytes_per_pixel in the qxl struct.
Eliminate assumptions in various places that the mode is always 32
bits. This almost, but not quite, makes it work with 16 bit modes. The
reason it doesn't work is that the QXL_BITMAP_FMT_16BIT assumes a 555
visual, but fb assumes an 565 visual.

But now at least it doesn't crash anymore.