Do the undamage handling in qxl_copy_n_to_n()
Previously it would do it in both copy_window() and copy_area(). The
problem with that, aside from duplication of a bit code, was that
sometimes the copy operations would be empty, which meant that no
damage was generated for the copy so that the pending damage region
was actually corresponding to some earlier operation, and so
discarding it caused rendering corruption.

This change moves the damage handling into qxl_copy_n_to_n() and only
does something when there is actually some copying going on.