config: move compiler flags from configure.ac to Makefile.am
CFLAGS is an automake defined variable that should not be set
by the module. It should not be AC_SUBST either, it already is.

Use AM_CFLAGS in Makefile.am. This will allow the user to override
the flags as they will be in the right order.

The -g option is already part of the default flags.

Signed-off-by: Gaetan Nadon <memsize@videotron.ca>