Make non default qxl modes available
Currently unless an xorg.conf with modelines is used a lot of the
qxl supported modes like 1920x1080 are not available, because the xorg
default modelines set does not have modelines for them.

This patch adds code to dynamically generate modelines for all modes
in the qxl mode list, bringing the xorg driver up to par with the windows
driver wrt supported resolutions.

Note that an xorg.conf specifying a large enough virtual screen size is
still needed for resolutions > 1024x768, but one no longer needs to add
modelines in it. This is fixed in my next patch.

This patch also adjusts a few (fake) clock limits to make all the
modes reported by the qxl device when compiled with a 16MB framebuffer
work.