Change default virtual size to match the highest available resolution
With the old default virtual size of 1024x768, using higher resolutions
is not possible without an xorg.conf. Since the default now a days is
to not have an xorg.conf, this is sort of unfortunate.

This patch makes these higher resolutions available, while keeping the
default resolution used when none is specified through xorg.conf at
1024x786, so that the spice client window won't be way too large for
smaller screens by default.

Note that when running inside a vm with a qxl device with a 8MB
framebuffer the code, for automatically setting the virtualsize to the
largest width and height seen in the resolution list, would lead to a too
large virtual size. The 8MB list has both 1920x1080 and 1600x1200
resolution which lead to a virtualsize of 1920x1200 which does not fit,
there is a special check for this situation, which maximizes the available
width in this case. This means that for using 1600x1200 on an 8MB device
an xorg.conf is still necessary.

This change does come at the prize of using more memory, but that
seems like a reasonable price to pay to give us parity wrt supported
resolutions with the windows driver. Also this is a must have to allow
the to be written Linux agent to change the guest resolution to match
the client machines one when running in auto fullscreen mode.

In the long run we should add support for resizing the fb on the fly to
match the resolution.