Make virtual resolution match actual resolution on startup
When setting a Virtual size in xorg.conf, currently there can be
a difference between the actual output resolution (say 1600x1200) and
the virtual resolution (for example 2048x1600). This causes parts
of the desktop too be unreachable because the cursor is bound to
the spicec window and thus one cannot pan over the virtual desktop.

Moreover other parts of the driver don't seem to like this and I've
seen several hangs / crashes which I think are related.