Fix the driver crashing qemu on 32 bits
When casting a 32bit pointer to a uint64 the following happens:
ptr -> int32 -> int64 -> uint64, so if the address is above
0x80000000 which is quite normal for mapped io, the int32 -> int64
cast causes sign extension, not good!

Also fix the printing of the memslots the memslot phys addresses
are always 64 bit, so tell printf to always read 64 bits, otherwise
we end up printing the higher 32 bits of the address as size on 32
bits.