xspice: init spice server, add main loop
Initialize a SpiceServer instance, and implement SpiceCoreInterface,
that is fd read, write notification and watchs (timers).

The SpiceServer instance creation is wrapped in xspice_get_spice_server to allow
access from the pointer and keyboard drivers introduced later.

The fd implementation is off because Xserver doesn't allow us to be
notified on write unblock, only read. Workaround is to poll.