Don't translate newly generated paccess region
The region passed to uxa_prepare_access() is in screen coordinates,
but the driver wants drawable coordinates. Hence we do a translation.

However, when the passed region is NULL, we generate the region
ourselves based on the full drawable extents. This region is already
in drawable space so shouldn't be translated.