xspice_keyboard_proc: fix arrow keys
Not sure yet why the regression with the arrow keys not producing the
right keysym, but it appears that the keycode is correct for up (as an
example), 111, but the keysym being produced was 0xff6c and not the
correct 0xff52. This boiled down to the rules of the default rmlvo being
"base" instead of "evdev". Providing an rmlvo parameter to
InitKeyboardDeviceStruct allows to override that. The chosen
 rules = "evdev"
 model = "pc105"
 layout = "us"
Is the default used by Xephyr from
xorg-x11-server-Xephyr-1.11.99.901-3.20120124.fc17.x86_64